method add_orders does completely all

+++ data/orders.py
@@ -6,6 +6,6 @@ class Order(SqlAlchemyBase):
     __tablename__ = 'orders'
 
     id = sqlalchemy.Column(sqlalchemy.Integer, primary_key=True)
-    weight = sqlalchemy.Column(sqlalchemy.Integer, nullable=True)
+    weight = sqlalchemy.Column(sqlalchemy.Float, nullable=True)
     region = sqlalchemy.Column(sqlalchemy.Integer, nullable=True)
     is_took = sqlalchemy.Column(sqlalchemy.Boolean, nullable=True)
